requirements.txt
streamlit
google-generative-ai
python-dotenv
langchain
langchain-community
PyPDF2
faiss-cpu
langchain-google-genai
dateparser
